сообщения ответов и ошибок вынесены в отдельный файл с константами
