+++ .gitignore
@@ -6,7 +6,7 @@ npm-debug.log*
 # Dependency directory
 node_modules
 .eslintrc/
-.env 
+.env
 
 # Optional npm cache directory
 .npm
