+++ routes/signup.js
@@ -1,5 +1,5 @@
 const router = require('express').Router();
-const { validatesingup } = require('../utils/constants');
+const { validatesingup } = require('../utils/validation');
 const { createUser } = require('../controllers/users');
 
 router.post('/', validatesingup, createUser);
