+++ utils/validation.js
@@ -1,0 +1,67 @@
+const { celebrate, Joi } = require('celebrate');
+
+const validateUserData = celebrate({
+  body: Joi.object().keys({
+    name: Joi.string().min(2).max(30).required(),
+    email: Joi.string().email().required(),
+  }),
+});
+
+const validateaddMovie = celebrate({
+  body: Joi.object().keys({
+    country: Joi.string().required(),
+    director: Joi.string().required(),
+    duration: Joi.number().required(),
+    year: Joi.string().required(),
+    description: Joi.string().required(),
+    image: Joi.string()
+      .required()
+      .regex(
+        /^https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_+.~#?&/=]*)$/,
+      ),
+    trailerLink: Joi.string()
+      .required()
+      .regex(
+        /^https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_+.~#?&/=]*)$/,
+      ),
+    nameRU: Joi.string().required(),
+    nameEN: Joi.string().required(),
+    thumbnail: Joi.string()
+      .required()
+      .regex(
+        /^https?:\/\/(?:www\.)?[-a-zA-Z0-9@:%._+~#=]{1,256}\.[a-zA-Z0-9()]{1,6}\b(?:[-a-zA-Z0-9()@:%_+.~#?&/=]*)$/,
+      ),
+    movieId: Joi.number().required(),
+  }),
+});
+
+const validatedeleteMovie = celebrate({
+  params: Joi.object().keys({
+    movieId: Joi.string().hex().required(),
+  }),
+});
+
+const validatesingin = celebrate({
+  body: Joi.object().keys({
+    email: Joi.string().required().email(),
+    password: Joi.string().required().min(5),
+  }).unknown(true),
+});
+
+const validatesingup = celebrate({
+  body: Joi.object()
+    .keys({
+      name: Joi.string().min(2).max(30),
+      email: Joi.string().required().email(),
+      password: Joi.string().required().min(5),
+    })
+    .unknown(true),
+});
+
+module.exports = {
+  validateUserData,
+  validateaddMovie,
+  validatedeleteMovie,
+  validatesingin,
+  validatesingup,
+};
